requirements.txt
phidata
python-dotenv
newspaper4k
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai